Add `const_assert_eq!` and `const_assert_ne!` macros
Leverages some idiosyncrasies of the Rust type system to define macros
which can be used to make compile-time assertions about const generic
parameters, working around the "can't use generic parameters from outer
item" error which would occur if one simply used `assert_eq`/`assert_ne`
