+++ src/lib.rs
@@ -155,7 +155,7 @@
 extern crate alloc;
 
 #[macro_use]
-mod nlimbs;
+mod macros;
 
 #[cfg(feature = "generic-array")]
 mod array;
